+++ SFML19_RoguelikeDungeon/SourceFiles/Floor/stair.cpp
@@ -7,7 +7,8 @@
 
 #include "Floor/stair.h"
 #include "Manager/texture_manager.h"
+#include <Floor/floor_object.h>
 
 Stair::Stair() : Floor_Object(Texture_Manager::stair) {}
 
-Stair::Stair(int x, int y) : Floor_Object(x, y, Texture_Manager::stair) {}
+Stair::Stair(float x, float y) : Floor_Object(x, y, Texture_Manager::stair) {}
